Add files via upload
Analysis of Unemployment, Inflation, and KRL Commuter Train Users in Indonesia

This analysis examines the relationship between unemployment, inflation, and the number of KRL Commuter Train users in Indonesia. Descriptive statistics will be used to analyze each factor individually. The aim is to understand the general trends and implications of these variables.